fix: 田野农谷反馈修复与企业权限/集团看板/同步异步化专项(含性能加固)
- 企业权限: role_menu_id 显式雪花主键修复保存失败; 失败整体回滚; tbmenu 定义随映射补齐
- 数据看板: 补集团总表端点 /hrmEmployee/departmentDetailGroupSummary(行政经理+看板权限)
- 同步考勤/获取审批: 异步化+按公司互斥锁+进度键按公司隔离+失败自动重试; 钉钉死循环快速失败; 单例共享状态清零; 排班分页节流; 同步日志保留期清理任务
- 员工管理: 用工性质/学历筛选、司龄计算修正(导入兜底改入职日期+Period 精确计算+存量修复 SQL)、合同到期动态状态、日期筛选解耦、表头固定、字段去重与字体对比度
- 基础设施: FTP/附件临时文件治理、动态调度线程池修复、备份快照清理、Redis mSet CME 修复、prod 日志降级
- 文档: docs 重组(modules/ 按模块拆分)与功能修改 Prompt 清单
- 并包含此前工作区存量变更(多租户小程序模块、SQL 基线、登录链路加固等)

+++ src/main/java/com/tianye/hrsystem/model/tbrolemenu.java
@@ -2,7 +2,6 @@ package com.tianye.hrsystem.model;
 
 import javax.persistence.*;
 import java.io.Serializable;
-import java.util.Date;
 
 @Entity
 @Table(name = "tbrolemenu")
@@ -10,17 +9,17 @@ public class tbrolemenu implements Serializable {
   @Id
   @GeneratedValue(strategy=GenerationType.IDENTITY)
   @Column(name = "role_menu_id")
-  private Integer roleMenuId;
+  private Long roleMenuId;
   @Column(name = "role_id")
   private Integer roleId;
   @Column(name = "menu_id")
   private Integer menuId;
 
-  public Integer getRoleMenuId() {
+  public Long getRoleMenuId() {
     return roleMenuId;
   }
 
-  public void setRoleMenuId(Integer roleMenuId) {
+  public void setRoleMenuId(Long roleMenuId) {
     this.roleMenuId = roleMenuId;
   }
